dashboard/pyproject.toml
[project]
name = "dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.2",
    "jinja2>=3.1.6",
    "pyjwt>=2.12.1",
]
